moves the compact card carousel into the molecules folder
